requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.9
django-bootstrap-v5==1.0.7
django-cors-headers==3.10.0
django-extensions==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
